+++ src/plone/restapi/serializer/user.py
@@ -1,15 +1,14 @@
 # -*- coding: utf-8 -*-
 from plone.restapi.interfaces import ISerializeToJson
+from plone.restapi.interfaces import ISerializeToJsonSummary
 from Products.CMFCore.interfaces._tools import IMemberData
 from zope.component import adapter
 from zope.component.hooks import getSite
 from zope.interface import implementer
-from zope.interface import Interface
+from zope.publisher.interfaces import IRequest
 
 
-@implementer(ISerializeToJson)
-@adapter(IMemberData, Interface)
-class SerializeUserToJson(object):
+class BaseSerializer(object):
 
     def __init__(self, context, request):
         self.context = context
@@ -39,3 +38,15 @@ class SerializeUserToJson(object):
             'location': user.getProperty('location'),
             'roles': roles,
         }
+
+
+@implementer(ISerializeToJson)
+@adapter(IMemberData, IRequest)
+class SerializeUserToJson(BaseSerializer):
+    pass
+
+
+@implementer(ISerializeToJsonSummary)
+@adapter(IMemberData, IRequest)
+class SerializeUserToJsonSummary(BaseSerializer):
+    pass