+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 New Features:
 
+- Manipulate group membership in the @groups endpoint.
+  [jaroel]
+
 - Expose a user's global roles though the @users endpoint.
   [jaroel]